Introduce hit all option
Sometimes we need to know whether the external call really happened,
as without the call system isn't fully functioning.

Currently where will be no sign of the absence of the call. This PR changes
that by introducing the `hit_all` option. It will raise an exception
if any of the previously recorded sessions weren't hit.